Changing Redirection after create and update to go back to the same location
When working with pages/layouts/snippets, especially while developing the first time around, after create and update redirects the user to the index page for pages/layouts/snippets.

This leads to a lot of unnecessary clicking as the user has to edit the page/layout/snippet again.

Also, if you are constant saving your work when you are adding page content, you go back to the page listing. And then have to come back in.

With these changes, you can continue editing till you click on "Back to xxx", e.g., "Back to Pages" or "Back to Layouts" or "Back to Snippets"

Also, changed references for model('categories') to model('category'). This way we use the CFWheels model function to actually call the model component, rather than going directly to the database.