Plugin zip files lowercased.
This is done as on *nix systems when they are unzipped, they create folders with mixed case names. And that creates folders that are different from unzipped plugin folders checked in the repo.

Ideal solution would be to check in all the unzipped plugin files with the correct casing as generated on a *nix system (which would also work on a Windows box), rather than changing the name of the plugin files.